// src/app/models/entities/core-order.model.ts
//BEGIN LICENSE BLOCK 
//Interneuron Terminus

//Copyright(C) 2024  Interneuron Limited

//This program is free software: you can redistribute it and/or modify
//it under the terms of the GNU General Public License as published by
//the Free Software Foundation, either version 3 of the License, or
//(at your option) any later version.

//This program is distributed in the hope that it will be useful,
//but WITHOUT ANY WARRANTY; without even the implied warranty of
//MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

//See the
//GNU General Public License for more details.

//You should have received a copy of the GNU General Public License
//along with this program.If not, see<http://www.gnu.org/licenses/>.
//END LICENSE BLOCK 
import { EntityBase } from './entity-base.model';
import { CoreNote } from './core-note.model';

export class CoreOrder extends EntityBase {
    person_id: string;
    order_id: string;
    observationdatetime: string;
    universalservicetext: string;
    resultstatus: string;
    category: string;
    ordertype: string;
    datetimeoftransaction: string;
    requesteddatetime: string;
    fillerordernumber: string;
    placerordernumber: string;
    orderingprovider: string;
    statuschangedatetime: string;
    specimentypecode: string;
    specimentypetext: string;
    specimendatetime: string;

    abnormalflagcount: number;
    totalviewcount: number;
    maxlastvieweddatetime: string;
    timeago: string;
}